requirements-lock.txt
asgiref==3.8.1
astroid==3.1.0
coverage==7.5.0
dill==0.3.8
Django==4.2.11
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-formtools==2.5.1
gunicorn==22.0.0
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.0
platformdirs==4.2.1
psycopg==3.1.18
psycopg-c==3.1.18
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
